Improved handling, Cleaned up code, Added support for injection keys,
Injection keys allow for variables to be used in inputted data. This way only one line needs to be updated for simple usage. An example is changing version numbers of deps. Where the number would need to be updated in the edit line and commit message. Now %version% can be used to update both at the same time via -key@version=0.1.2